Add "Archival Delay" configuration option.
Don't download episodes in "initial" quality within a configurable number of days from the air date, IF "archival" quality is specified for a show AND other episodes have already been downloaded in archival quality.

This prevents multiple downloads when the archival quality is frequently released a short time after the initial quality. It won't prevent immediate download of episodes in initial quality for shows that are not released in archival quality.

You should now be able to specify a single custom quality setting (e.g. initial: SD DVD, HDTV; archival: 720p WEB-DL) that works for all shows without frequent duplicate downloads, and without having to know in advance which shows are going to be released in archival quality.